split user_accounts models into separate files
