+++ user_accounts/exceptions.py
@@ -8,3 +8,15 @@ class UnacceptedInviteError(Exception):
     '''The user has not yet accepted this invite
     '''
     pass
+
+
+class UndefinedResourceAccessError(Exception):
+    '''Someone should not have access to this resource
+    '''
+    pass
+
+
+class NoEmailsForOrgError(Exception):
+    '''There is no email contact available for this organization
+    '''
+    pass
